Apply DPI etc. based on application loading the resources...
... and not based on the resources' package. Makes no difference for
applications that load only their own resources. However, e.g.
notifications are loaded by the SystemUI and they are also displayed in
the context of the SystemUI. So for consistency, it is better to use the
same DPI for all parts of the application - including the parts loaded
from foreign packages.

Locale settings are not affected, they are still applied for the loaded
package.