Add setting to mute audio.
This intercepts the several Audio java client APIs running in the context of the target packages and suppresses the play and resume actions.
In the case of MediaPlayer, which can also handle video, a check is made for each track if a video surface was set; if so, the play action proceeds but the volume is set to 0 upfront, allowing the video to play but without sound.

Not all applications will be completely covered by this, since native APIs - such as OpenSL ES - can be used and are not possible to be intercepted with Xposed.
On the side of the audio service it is also not possible to perform the interception, since it is completely implemented in native code.