Adjust resources replacement to match XposedBridge v52.
Due to some internal changes in the resources replacement on the framework, the hook
that overrides resource configuration settings takes place before the object is
replaced by XResources.
The mechanism to look up which package the Resources belong to was updated to
use a new method introduced on the framework for this purpose.