Do not require reboot to apply "No fullscreen IME" setting.
The hook where this check is made runs in the context of the IME process and not on the target app's nor the system_server.
For that reason, it doesn't receive the broadcast notification of settings changed nor does it get reloaded when the target app is killed.
With whis change no system restart is required, nor does the target app need to be killed.